Use native PIO interface instead Arduino functions
I realize this looks like black magic to anyone who isn't familiar with
this stuff. I promise I'm going to explain all of it in a future blog
post. Until then, you read up on it in the reference documentation for
the Atmel SAM3X8E, chapter 31 and more specifically, chapter 31.7.

The reference documentation is available at:
http://www.atmel.com/devices/SAM3X8E.aspx